requirements-hf.txt
gradio==4.19.0
huggingface_hub==0.20.0
openai-whisper==20231117
librosa==0.10.1
numpy==1.24.3
scipy==1.11.4
numba==0.58.1
